analisis_numerico/requirements.txt
streamlit>=1.32.0
numpy>=1.26.0
scipy>=1.12.0
sympy>=1.12
matplotlib>=3.8.0
opencv-python-headless>=4.9.0
pytesseract>=0.3.10
Pillow>=10.2.0
pandas>=2.2.0
